+++ cognite_toolkit/_cdf_tk/builders/_base.py
@@ -1,95 +1,45 @@
 import difflib
-import re
-from collections import defaultdict
-from collections.abc import Hashable, Sequence, Iterable
-from dataclasses import dataclass, field
+from abc import abstractmethod, ABC
+from collections.abc import Iterable
 from pathlib import Path
-from typing import Any, ClassVar
+from typing import ClassVar
 
-import yaml
-from cognite.client.data_classes import Source
-from pandas.core.base import NoNewAttributesMixin
-
-from cognite_toolkit._cdf_tk.constants import INDEX_PATTERN, TEMPLATE_VARS_FILE_SUFFIXES
+from cognite_toolkit._cdf_tk.constants import INDEX_PATTERN, YAML_SUFFIX
 from cognite_toolkit._cdf_tk.data_classes import (
-    BuildVariables,
-    BuiltResource,
-    BuiltResourceList,
     ModuleLocation,
-    SourceLocationEager, SourceLocation,
+    BuildDestinationFile, BuildSourceFile, BuiltResourceList,
 )
 from cognite_toolkit._cdf_tk.exceptions import (
     AmbiguousResourceFileError,
-    ToolkitYAMLFormatError,
 )
 from cognite_toolkit._cdf_tk.loaders import (
     LOADER_BY_FOLDER_NAME,
     GroupLoader,
     Loader,
-    RawDatabaseLoader,
-    RawTableLoader,
-    ResourceLoader,
+    RawTableLoader, ResourceLoader,
 )
 from cognite_toolkit._cdf_tk.tk_warnings import (
-    ToolkitNotSupportedWarning,
-    ToolkitWarning,
-    WarningList,
+    ToolkitNotSupportedWarning, WarningList, ToolkitWarning,
 )
 from cognite_toolkit._cdf_tk.tk_warnings.fileread import (
-    DuplicatedItemWarning,
-    FileReadWarning,
-    MissingRequiredIdentifierWarning,
     UnknownResourceTypeWarning,
-    UnresolvedVariableWarning,
 )
 from cognite_toolkit._cdf_tk.utils import (
-    calculate_str_or_file_hash,
     humanize_collection,
-    quote_int_value_by_key_in_yaml,
-    safe_read,
-    safe_write,
 )
-from cognite_toolkit._cdf_tk.validation import validate_data_set_is_set, validate_resource_yaml
 
-@dataclass
-class BuildSourceFile:
-    source: SourceLocation
-    content: str
-    loaded: list[dict[str, Any]] | dict[str, Any] | None = None
-    warnings: WarningList[FileReadWarning] = field(default_factory=WarningList[FileReadWarning])
 
-@dataclass
-class BuildDestinationFile:
-    path: Path
-    loaded: list[dict[str, Any]] | dict[str, Any]
-    loader: type[ResourceLoader]
-    source: SourceLocation
-    extra_sources: list[SourceLocation] | None
-
-class Builder:
+class Builder(ABC):
     _resource_folder: ClassVar[str | None] = None
 
     def __init__(
         self,
         build_dir: Path,
-        module_names_by_variable_key: dict[str, list[str]],
-        silent: bool,
-        verbose: bool,
         resource_folder: str | None = None,
     ):
         self.build_dir = build_dir
-        self.silent = silent
-        self.warning_list = WarningList[ToolkitWarning]()
-        self.verbose = verbose
-
         self.resource_counter = 0
-        self._warning_index = 0
         self.index_by_filepath_stem: dict[Path, int] = {}
-        self.ids_by_resource_type: dict[type[ResourceLoader], dict[Hashable, Path]] = defaultdict(dict)
-        self.dependencies_by_required: dict[tuple[type[ResourceLoader], Hashable], list[tuple[Hashable, Path]]] = (
-            defaultdict(list)
-        )
-        self._module_names_by_variable_key = module_names_by_variable_key
         if self._resource_folder is not None:
             self.resource_folder = self._resource_folder
         elif resource_folder is not None:
@@ -97,86 +47,15 @@ class Builder:
         else:
             raise ValueError("Either _resource_folder or resource_folder must be set.")
 
-    @property
-    def print_warning(self) -> bool:
-        return not self.silent
-
-    def warn(self, warning: ToolkitWarning) -> None:
-        self.warning_list.append(warning)
-        if not self.silent:
-            warning.print_warning()
-
-    def console(self, message: str, prefix: str = "[bold green]INFO:[/] ") -> None:
-        if not self.silent:
-            print(f"{prefix}{message}")
-
-    def build_resource_folder(
-        self, resource_files: Sequence[Path], module_variables: BuildVariables, module: ModuleLocation
-    ) -> BuiltResourceList[Hashable]:
-        self._warning_index = len(self.warning_list)
-        built_resource_list = BuiltResourceList[Hashable]()
-
-        source_files = self._replace_variables(resource_files, module_variables, module.dir)
-
-        for destination_file in self.build(source_files, module):
-            file_warnings, identifiers_kind_pairs = self.validate(destination_file.loaded, destination_file.source.path, destination_file.path)
-
-            if file_warnings:
-                self.warning_list.extend(file_warnings)
-                # Here we do not use the self.warn method as we want to print the warnings as a group.
-                if self.print_warning:
-                    print(str(file_warnings))
-
-            built = BuiltResourceList(
-                [BuiltResource(identifier, destination_file.source, kind, destination_file.path, extra_sources=destination_file.extra_sources) for identifier, kind in
-                 identifiers_kind_pairs]
-            )
-            built_resource_list.extend(built)
-
-        return built_resource_list
-
+    @abstractmethod
     def build(self, source_files: list[BuildSourceFile], module: ModuleLocation) -> Iterable[BuildDestinationFile]:
-        ...
+        raise NotImplementedError()
 
-    def _replace_variables(self, resource_files: Sequence[Path], variables: BuildVariables, module_dir: Path) -> list[BuildSourceFile]:
-        source_files: list[BuildSourceFile] = []
+    def validate_folder(self, built_resources: BuiltResourceList, module: ModuleLocation) -> WarningList[ToolkitWarning]:
+        """This can be overridden to add additional validation for the built resources."""
+        return WarningList[ToolkitWarning]()
 
-        for source_path in resource_files:
-            content = safe_read(source_path)
-            source = SourceLocationEager(source_path, calculate_str_or_file_hash(content, shorten=True))
-
-            if source_path.suffix.lower() not in TEMPLATE_VARS_FILE_SUFFIXES:
-                source_files.append(BuildSourceFile(source, content, None, WarningList[FileReadWarning]()))
-                continue
-
-            content = variables.replace(content, source_path.suffix)
-
-            warnings = self._check_variables_replaced(content, module_dir, source_path)
-
-            if source_path.suffix not in {".yaml", ".yml"}:
-                source_files.append(BuildSourceFile(source, content, None, warnings))
-                continue
-
-            # Ensure that all keys that are version gets read as strings.
-            # This is required by DataModels, Views, and Transformations that reference DataModels and Views.
-            content = quote_int_value_by_key_in_yaml(content, key="version")
-            try:
-                parsed = yaml.safe_load(content)
-            except yaml.YAMLError as e:
-                if self.print_warning:
-                    print(str(warnings))
-                raise ToolkitYAMLFormatError(
-                    f"YAML validation error for {source_path.name} after substituting config variables: {e}"
-                )
-
-            source_files.append(BuildSourceFile(source, content, parsed, warnings))
-
-        return source_files
-
-
-    def last_build_warnings(self) -> WarningList[ToolkitWarning]:
-        return self.warning_list[self._warning_index :]
-
+    # Helper methods
     def _create_destination_path(self, source_path: Path, module_dir: Path) -> Path:
         """Creates the filepath in the build directory for the given source path.
 
@@ -205,84 +84,8 @@ class Builder:
         destination_path.parent.mkdir(parents=True, exist_ok=True)
         return destination_path
 
-    def validate(
-        self,
-        parsed: dict[str, Any] | list[dict[str, Any]],
-        source_path: Path,
-        destination: Path,
-    ) -> tuple[WarningList[FileReadWarning], list[tuple[Hashable, str]]]:
-        warning_list = WarningList[FileReadWarning]()
 
-        loader = self._get_loader(self.resource_folder, destination, source_path)
-        if loader is None or not issubclass(loader, ResourceLoader):
-            return warning_list, []
-
-        is_dict_item = isinstance(parsed, dict)
-        items = [parsed] if is_dict_item else parsed
-
-        identifier_kind_pairs: list[tuple[Hashable, str]] = []
-        for no, item in enumerate(items, 1):
-            element_no = None if is_dict_item else no
-
-            identifier: Any | None = None
-            # Raw Tables and Raw Databases can have different loaders in the same file.
-            item_loader = loader
-            try:
-                identifier = item_loader.get_id(item)
-            except KeyError as error:
-                if loader is RawTableLoader:
-                    try:
-                        identifier = RawDatabaseLoader.get_id(item)
-                        item_loader = RawDatabaseLoader
-                    except KeyError:
-                        warning_list.append(
-                            MissingRequiredIdentifierWarning(source_path, element_no, tuple(), error.args)
-                        )
-                else:
-                    warning_list.append(MissingRequiredIdentifierWarning(source_path, element_no, tuple(), error.args))
-
-            if identifier:
-                identifier_kind_pairs.append((identifier, item_loader.kind))
-                if first_seen := self.ids_by_resource_type[item_loader].get(identifier):
-                    warning_list.append(DuplicatedItemWarning(source_path, identifier, first_seen))
-                else:
-                    self.ids_by_resource_type[item_loader][identifier] = source_path
-
-                for dependency in item_loader.get_dependent_items(item):
-                    self.dependencies_by_required[dependency].append((identifier, source_path))
-
-            api_spec = item_loader.safe_get_write_cls_parameter_spec()
-            if api_spec is not None:
-                resource_warnings = validate_resource_yaml(parsed, api_spec, source_path, element_no)
-                warning_list.extend(resource_warnings)
-
-            data_set_warnings = validate_data_set_is_set(items, loader.resource_cls, source_path)
-            warning_list.extend(data_set_warnings)
-
-        return warning_list, identifier_kind_pairs
-
-    def _check_variables_replaced(self, content: str, module: str, source_path: Path) -> WarningList[FileReadWarning]:
-        all_unmatched = re.findall(pattern=r"\{\{.*?\}\}", string=content)
-        warning_list = WarningList[FileReadWarning]()
-        for unmatched in all_unmatched:
-            warning_list.append(UnresolvedVariableWarning(source_path, unmatched))
-            variable = unmatched[2:-2]
-            if module_names := self._module_names_by_variable_key.get(variable):
-                module_str = (
-                    f"{module_names[0]!r}"
-                    if len(module_names) == 1
-                    else (", ".join(module_names[:-1]) + f" or {module_names[-1]}")
-                )
-                self.console(
-                    f"The variables in 'config.[ENV].yaml' need to be organised in a tree structure following"
-                    f"\n    the folder structure of the modules, but can also be moved up the config hierarchy to be shared between modules."
-                    f"\n    The variable {variable!r} is defined in the variable section{'s' if len(module_names) > 1 else ''} {module_str}."
-                    f"\n    Check that {'these paths reflect' if len(module_names) > 1 else 'this path reflects'} the location of {module}.",
-                    prefix="    [bold green]Hint:[/] ",
-                )
-        return warning_list
-
-    def _get_loader(self, resource_folder: str, destination: Path, source_path: Path) -> type[Loader] | None:
+    def _get_loader(self, resource_folder: str, destination: Path, source_path: Path) -> type[ResourceLoader] | None:
         folder_loaders = LOADER_BY_FOLDER_NAME.get(resource_folder, [])
         if not folder_loaders:
             self.warn(
@@ -328,3 +131,23 @@ class Builder:
             )
 
         return loaders[0]
+
+
+class DefaultBuilder(Builder):
+    def build(self, source_files: list[BuildSourceFile], module: ModuleLocation) -> Iterable[BuildDestinationFile]:
+        for source_file in source_files:
+            if source_file.source.path.suffix.lower() not in YAML_SUFFIX:
+                continue
+            destination_path = self._create_destination_path(source_file.source.path, module.dir)
+            loader = self._get_loader(self.resource_folder, destination_path, source_file.source.path)
+            if loader is None:
+                continue
+
+            destination = BuildDestinationFile(
+                path=destination_path,
+                loaded=source_file.loaded,
+                loader=loader,
+                source=source_file.source,
+                extra_sources=None,
+            )
+            yield destination
